Add a joke in the commit
Santa and a friend are sitting in a cinema. Just before the break they see a cactus and in some distance a cowboy. During the break the friend says to santa:"I bet the cowboy will ride into the cactus."
Santa answers: "I do not believe that."
They agree that the loser invites the winner to a bottle of wine after the film. It turns out that the friend wins. So after the film they drink together the bottle of wine in a restaurant near the cinema.
Then the friend says: "I must confess that the bet was not fair. I saw the film for the second time."
Then Santa replies: "And I saw it for the fourth time, but I did not think that this fool rides into the cactus again."